pyproject.toml
[tool.poetry]
name = "raspar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
schematics = "^2.1.1"
bs4 = "^0.0.1"
requests = "^2.26.0"
persist-queue = "^0.7.0"
pybloomfilter = "^1.0"
fake-useragent = "^0.1.11"
python-dateutil = "^2.8.2"
validators = "^0.18.2"
persistQueue = "^0.1.6"
furl = "^2.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
